feat: complete security audit remediation (Sprints 1-5)
Implements all 99 remediation items from the security audit across
5 sprints: security blockers, correctness & stability, hardening,
quality & testing, and enterprise readiness.

Key changes:
- Fix all critical/high security findings (RCE, injection, secrets exposure)
- Add CSRF protection, TLS hardening, input validation, rate limiting
- Add JPA @Version optimistic locking to concurrent entities
- Add Flyway database migrations (replacing hibernate ddl-auto)
- Add X.509 extensions (SKI, AKI, SAN, EKU) to all certificate paths
- Add CRL generation, CSR validation, CA expiry checks
- Add split-brain resolution and graceful transfer draining for clustering
- Add NetworkPolicy Helm templates for pod-level network isolation
- Fix all unit test failures (CSRF tokens, mock setup, socket timeouts)
- 1710 tests passing across all modules (0 failures, 0 errors)

+++ pesitwizard-pesit/src/main/java/com/pesitwizard/fpdu/ParameterValueType.java
@@ -17,7 +17,7 @@ public enum ParameterValueType {
     N {
         @Override
         public String renderValue(ParameterValue value) {
-            int n = 0;
+            long n = 0;
             for (byte b : value.getValue()) {
                 n = (n << 8) | (b & 0xFF);
             }
@@ -28,7 +28,7 @@ public enum ParameterValueType {
     S {
         @Override
         public String renderValue(ParameterValue value) {
-            int n = 0;
+            long n = 0;
             for (byte b : value.getValue()) {
                 n = (n << 8) | (b & 0xFF);
             }
@@ -39,11 +39,11 @@ public enum ParameterValueType {
     M {
         @Override
         public String renderValue(ParameterValue value) {
-            int n = 0;
+            long n = 0;
             for (byte b : value.getValue()) {
                 n = (n << 8) | (b & 0xFF);
             }
-            return Integer.toBinaryString(n);
+            return Long.toBinaryString(n);
         }
     },
     // Date/Heure / DateTime
